ax88179_178a: fix bonding failure
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-514.el7
commit-author Ian Morgan <imorgan@primordial.ca>
commit 95ff88688781db2f64042e69bd499e518bbb36e5

The following patch fixes a bug which causes the ax88179_178a driver to be
incapable of being added to a bond.

When I brought up the issue with the bonding maintainers, they indicated
that the real problem was with the NIC driver which must return zero for
success (of setting the MAC address). I see that several other NIC drivers
follow that pattern by either simply always returing zero, or by passing
through a negative (error) result while rewriting any positive return code
to zero. With that same philisophy applied to the ax88179_178a driver, it
allows it to work correctly with the bonding driver.

I believe this is suitable for queuing in -stable, as it's a small, simple,
and obvious fix that corrects a defect with no other known workaround.

This patch is against vanilla 3.17(.0).

	Signed-off-by: Ian Morgan <imorgan@primordial.ca>

 drivers/net/usb/ax88179_178a.c |    7 ++++++-
 1 file changed, 6 insertions(+), 1 deletion(-)
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 95ff88688781db2f64042e69bd499e518bbb36e5)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>