dm snapshot: don't invalidate on-disk image on snapshot write overflow
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-514.el7
commit-author Mikulas Patocka <mpatocka@redhat.com>
commit 76c44f6d80e151d230844db7ffc058ac21b9e3f2

When the snapshot overflows because of a write to the origin, the on-disk
image has to be invalidated.  However, when the snapshot overflows because
of a write to the snapshot, the on-disk image doesn't have to be
invalidated.  Change the behavior so that the on-disk image is not
invalidated in this case.

When the snapshot overflows, the variable snapshot_overflowed is set.
All writes to the snapshot are disallowed to minimize filesystem
corruption - this condition is cleared when the snapshot is deactivated
and activated.

The user can extend the overflowed snapshot, deactivate and activate it
again, run fsck (if journaling filesystem is not used) mount it and
recover the data.

	Signed-off-by: Mikulas Patocka <mpatocka@redhat.com>
	Signed-off-by: Mike Snitzer <snitzer@redhat.com>
(cherry picked from commit 76c44f6d80e151d230844db7ffc058ac21b9e3f2)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>