dm thin: restore requested 'error_if_no_space' setting on OODS to WRITE transition
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-514.el7
commit-author Mike Snitzer <snitzer@redhat.com>
commit 172c238612ebf81cabccc86b788c9209af591f61

A thin-pool that is in out-of-data-space (OODS) mode may transition back
to write mode -- without the admin adding more space to the thin-pool --
if/when blocks are released (either by deleting thin devices or
discarding provisioned blocks).

But as part of the thin-pool's earlier transition to out-of-data-space
mode the thin-pool may have set the 'error_if_no_space' flag to true if
the no_space_timeout expires without more space having been made
available.  That implementation detail, of changing the pool's
error_if_no_space setting, needs to be reset back to the default that
the user specified when the thin-pool's table was loaded.

Otherwise we'll drop the user requested behaviour on the floor when this
out-of-data-space to write mode transition occurs.

	Reported-by: Vivek Goyal <vgoyal@redhat.com>
	Signed-off-by: Mike Snitzer <snitzer@redhat.com>
	Acked-by: Joe Thornber <ejt@redhat.com>
Fixes: 2c43fd26e4 ("dm thin: fix missing out-of-data-space to write mode transition if blocks are released")
	Cc: stable@vger.kernel.org
(cherry picked from commit 172c238612ebf81cabccc86b788c9209af591f61)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>