SUNRPC: never enqueue a ->rq_cong request on ->sending
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-514.el7
commit-author Neil Brown <neilb@suse.de>
commit 298073181112a6ab6c30fe7971b99de968daf81e

If the sending queue has a task without ->rq_cong set at the front,
and then a number of tasks with ->rq_cong set such that they use
the entire congestion window, then the queue deadlocks.  The first
entry cannot be processed until later entries complete.

This scenario has been seen with a client using UDP to access a server,
and the network connection breaking for a period of time - it doesn't
recover.

It never really makes sense for an ->rq_cong request to be on the ->sending
queue, but it can happen when a request is being retried, and finds
the transport if locked (XPRT_LOCKED).  In this case we simple call
__xprt_put_cong() and the deadlock goes away.

	Signed-off-by: NeilBrown <neilb@suse.de>
	Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
(cherry picked from commit 298073181112a6ab6c30fe7971b99de968daf81e)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>