Btrfs: only adjust outstanding_extents when we do a short write
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-514.el7
commit-author Josef Bacik <jbacik@fb.com>
commit 3e05bde8c3c2dd761da4d52944a087907955a53c

We have this weird dance where we always inc outstanding_extents when we do a
O_DIRECT write, even if we allocate the entire range.  To get around this we
also drop the metadata space if we successfully write.  This is an unnecessary
dance, we only need to jack up outstanding_extents if we don't satisfy the
entire range request in get_blocks_direct, otherwise we are good using our
original reservation.  So drop the unconditional inc and the drop of the
metadata space that we have for the unconditional inc.  Thanks,

	Signed-off-by: Josef Bacik <jbacik@fb.com>
	Reviewed-by: Liu Bo <bo.li.liu@oracle.com>
	Signed-off-by: Chris Mason <clm@fb.com>
(cherry picked from commit 3e05bde8c3c2dd761da4d52944a087907955a53c)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>