btrfs: qgroup: do a reservation in a higher level.
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-514.el7
Rebuild_CHGLOG: - [fs] btrfs: qgroup: do a reservation in a higher level (Bill O'Donnell) [1298680]
Rebuild_FUZZ: 98.99%
commit-author Dongsheng Yang <yangds.fnst@cn.fujitsu.com>
commit e2d1f92399afb6ec518b68867ed10db2585b283a

There are two problems in qgroup:

a). The PAGE_CACHE is 4K, even when we are writing a data of 1K,
qgroup will reserve a 4K size. It will cause the last 3K in a qgroup
is not available to user.

b). When user is writing a inline data, qgroup will not reserve it,
it means this is a window we can exceed the limit of a qgroup.

The main idea of this patch is reserving the data size of write_bytes
rather than the reserve_bytes. It means qgroup will not care about
the data size btrfs will reserve for user, but only care about the
data size user is going to write. Then reserve it when user want to
write and release it in transaction committed.

In this way, qgroup can be released from the complex procedure in
btrfs and only do the reserve when user want to write and account
when the data is written in commit_transaction().

	Signed-off-by: Dongsheng Yang <yangds.fnst@cn.fujitsu.com>
	Signed-off-by: Chris Mason <clm@fb.com>
(cherry picked from commit e2d1f92399afb6ec518b68867ed10db2585b283a)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>