btrfs: make root id query unprivileged
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-514.el7
commit-author David Sterba <dsterba@suse.cz>
commit 01b810b889d557257580970e1a7ba9c85b54766b

The INO_LOOKUP ioctl can lookup path for a given inode number and is
thus restricted. As a sideefect it can find the root id of the
containing subvolume and we're using this int the 'btrfs inspect rootid'
command.

The restriction is unnecessary in case we set the ioctl args
 args::treeid    = 0
 args::objectid  = 256 (BTRFS_FIRST_FREE_OBJECTID)

Then the path will be empty and the treeid is filled with the root id of
the inode on which the ioctl is called. This behaviour is unchanged,
after the root restriction is removed.

	Signed-off-by: David Sterba <dsterba@suse.cz>
	Signed-off-by: Chris Mason <clm@fb.com>
(cherry picked from commit 01b810b889d557257580970e1a7ba9c85b54766b)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>